Rename unit tests for Windows regional date-time formatting
